fix(notice): a_value 를 BigInteger 로 + 기초금액 배치 건별 격리 (#83)
운영 실행에서 기초금액 수집이 죽었다:

    NumericValueOutOfRange: integer out of range
    a_value=2168128646  (기초금액 298억 공고)

`Notice.a_value` 가 Integer(int4 상한 2,147,483,647) 였다. A값은 대형 공사에서
이 값을 넘는다. `MockBid.price` 에서 겪은 것과 같은 함정이다.
`net_cost`(순공사원가)도 같은 이유로 함께 넓힌다. 마이그 b6f1d3a48c27.

더 나쁜 건 **한 건이 하루치를 통째로 날린 것**이다. `apply_to_notice` 는
세션에 얹기만 하고 실제 오류는 `commit()` 에서 나므로, 건별 try/except 가
있어도 소용이 없었다(287건 파싱·반영 후 커밋에서 전량 롤백).

→ `_apply_batch`: 일괄 커밋이 실패하면 **건별로 격리 재시도**한다. 실패는 그
한 건에 갇히고 `failed` 카운터에 남는다(nurture.py 의 "실패는 1건에 갇힌다"
와 같은 원칙).

테스트 787건 통과(+2), ruff 통과, alembic head 단일(b6f1d3a48c27).
- 일괄 커밋 실패 시 건별 재시도로 전량 살아나는지
- a_value·net_cost 컬럼이 BigInteger 인지 (회귀 방지선)

+++ backend/app/db/models.py
@@ -172,8 +172,11 @@ class Notice(Base):
     prdprc_range_end = Column(Float)           # rsrvtnPrceRngEndRate (예: +2)
 
     # Calculator Fields
-    a_value = Column(Integer, default=0)
-    net_cost = Column(Integer, default=0)
+    # ⚠️ BigInteger 필수 — A값은 대형 공사에서 21.4억(int4 상한)을 넘는다.
+    # 실측: 기초금액 298억 공고의 A값 2,168,128,646 (2026-08-05, 배치가
+    # NumericValueOutOfRange 로 죽고 그날치 커밋이 통째로 롤백됐다).
+    a_value = Column(BigInteger, default=0)
+    net_cost = Column(BigInteger, default=0)
     # A값 출처 — tier0(공고 API) | tier1(익스텐션) | tier2(첨부파싱) | none.
     # tier0 는 조달청이 공고에 실어 주는 값이라 가장 신뢰도가 높다.
     a_value_source = Column(String(10))
